codes/python-graphql-fetching/requirements.txt
requests
pandas
boto3
google-cloud-datastore
sentry_sdk
